# backend/openapi/deadmans.v1.yaml
openapi: 3.0.3
info:
  title: Dead-Mans API
  version: v1
  description: >
    Canonical transport contract for Dead-Mans: HTTP paths under /api and SignalR hubs
    documented in x-signalr. Regenerate frontend contracts after changes.
servers:
  - url: /api
paths:
  /game:
    get:
      operationId: getCurrentGameBoard
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Active, ready (registration), or latest finished game board
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameBoardSnapshotDto'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No active, ready, or finished game available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /game/team-queue:
    get:
      operationId: getCurrentGameTeamQueue
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Confirmed team queue for the current active or ready game
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GameTeamQueueItemDto'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /game/active-team:
    put:
      operationId: setActiveGameTeam
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetActiveGameTeamRequestDto'
      responses:
        '204':
          description: Active team selected or cleared
        '400':
          description: Invalid team id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing moderator/admin role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Active game or team not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Team cannot become active
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /game/cells/{cellId}/open:
    post:
      operationId: openGameCell
      security:
        - cookieAuth: []
      parameters:
        - name: cellId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Card opened or already open
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing admin role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Active team must be selected first
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  responses:
    InternalServerError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: dm_auth
  schemas:
    GameBoardCellMediaDto:
      type: object
      required: [url]
      properties:
        url:
          type: string
    GameBoardCellDto:
      type: object
      required: [id, row, col, cellType, cost, state, media]
      properties:
        id:
          type: string
        row:
          type: integer
        col:
          type: integer
        cellType:
          type: string
        title:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        cost:
          type: integer
        state:
          type: string
          enum: [closed, open]
        media:
          type: array
          items:
            $ref: '#/components/schemas/GameBoardCellMediaDto'
...
    GameBoardSnapshotDto:
      type: object
      required:
        - gameId
        - title
        - status
        - version
        - rows
        - cols
        - rowLabels
        - colLabels
        - cells
        - enabledModifierIds
        - activeModifiers
      properties:
        gameId:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [ready, active, finished]
        version:
          type: integer
        rows:
          type: integer
        cols:
          type: integer
        rowLabels:
          type: array
          items:
            type: string
        colLabels:
          type: array
          items:
            type: string
        cells:
          type: array
          items:
            $ref: '#/components/schemas/GameBoardCellDto'
        enabledModifierIds:
          type: array
          items:
            type: string
            format: uuid
        enabledQuestionIds:
          type: array
          items:
            type: string
            format: uuid
        activeModifiers:
          type: array
          items:
            $ref: '#/components/schemas/GameModifierActivationDto'
        activeTeamId:
          type: string
          format: uuid
          nullable: true
    SetActiveGameTeamRequestDto:
      type: object
      properties:
        teamId:
          type: string
          format: uuid
          nullable: true
    GameTeamQueueParticipantDto:
      type: object
      required: [userId, displayName]
      properties:
        userId:
          type: string
          format: uuid
        displayName:
          type: string
    GameTeamQueueItemDto:
      type: object
      required: [teamId, teamSlotIndex, participants]
      properties:
        teamId:
          type: string
          format: uuid
        teamSlotIndex:
          type: integer
        participants:
          type: array
          items:
            $ref: '#/components/schemas/GameTeamQueueParticipantDto'
...
    GameModifierActivationDto:
      type: object
      required: [modifierId, activatedByUserId, activatedAtUtc]
      properties:
        modifierId:
          type: string
          format: uuid
        activatedByUserId:
          type: string
          format: uuid
        activatedAtUtc:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
          nullable: true
          description: Stable machine-readable error code.
          enum:
            - game_board.not_found
            - game_board.cell_not_found
            - game_board.active_team_required
            - game_board.active_team_no_active_game
            - game_board.active_team_not_found
            - game_board.active_team_not_confirmed
            - game_board.active_team_has_no_active_members
            - game_setup.no_draft
            - game_setup.draft_exists
            - game_setup.invalid_title
            - game_setup.invalid_save_request
            - game_setup.cell_not_found
            - game_setup.cell_media_not_found
            - game_setup.invalid_cell_media_upload
            - game_setup.stale_version
            - game_lifecycle.draft_not_found
            - game_lifecycle.ready_already_exists
            - game_lifecycle.active_already_exists
            - game_lifecycle.game_not_ready
            - game_lifecycle.game_not_active
            - game_lifecycle.registration_slots_required
            - game_lifecycle.invalid_team_size_limits
            - game_lifecycle.no_confirmed_teams
            - game_lifecycle.unconfirmed_teams
            - game_lifecycle.pending_invitations
            - game_lifecycle.pending_disband_requests
            - game_lifecycle.invalid_confirmed_team_roster
            - game_lifecycle.operation_failed
            - game_lifecycle.draft_delete_not_allowed
            - game_lifecycle.game_not_found
            - game_common.unexpected_server_error
            - game_common.too_many_requests
            - game_registration.not_open
            - game_registration.no_slots
            - game_registration.already_on_team
            - game_registration.team_not_found
            - game_registration.team_not_joinable
            - game_registration.not_team_member
            - game_registration.invitation_invalid
            - game_registration.slot_not_found
            - game_registration.slot_not_available
            - game_registration.user_not_found
            - game_registration.pending_invitation
            - game_registration.pending_outgoing_invitation
            - game_registration.team_invite_not_allowed
            - game_registration.team_active_in_game
            - game_registration.operation_failed
            - game_modifier.game_not_active
            - game_modifier.not_enabled
            - game_modifier.conflict_active
            - game_modifier.limit_reached
            - game_modifier.user_not_resolved
            - game_modifier.invalid_request
            - game_modifier.not_found
            - game_card_run.no_active_game
            - game_card_run.cell_not_found
            - game_card_run.cell_not_open
            - game_card_run.team_not_found
            - game_card_run.team_not_confirmed
            - game_card_run.team_has_no_active_members
            - game_card_run.already_in_progress
            - game_card_run.invalid_request
            - game_card_run.not_found
            - game_card_run.not_in_progress
            - game_card_run.modifier_result_not_found
            - game_question.invalid_request
            - game_question.duplicate_code
            - game_question.not_found
            - game_question.category_not_found
            - game_question.category_not_empty
            - game_question.category_protected
            - game_question.import_invalid_fields
            - game_question.import_duplicate_code_in_file
            - game_question.import_category_unresolved
            - game_question.import_duplicate_code_existing
            - game_question.no_active_game
            - game_question.no_available_questions
            - game_question.round_not_found
            - game_question.round_not_pending
            - game_question.manual_award_player_not_found
            - game_question.manual_award_invalid_points
        requestId:
          type: string
          nullable: true
          description: Server request correlation identifier for diagnostics.
...